Added repositories endpoint
Necessary to encapsulate the logic for making an API request to the
repository endpoint for users and organizations.

RSpec support fixtures are being used to represent user and
organization repositories. In both cases the repositories are of the
same name (i.e. `test`) but own is owned by me (user) and the other is
private (organization).