Added custom code quality configurations
Necessary to ignore longer parameters lists for the various endpoints
due to GitHub's API being difficult to deal with in terms of nested
routes (this effects both Reek and RuboCop).

The Reek data clump issues are related to the above except for the API
client which is mostly an issue with the HTTP gem.

The RuboCop metric issues are largely due to dealing with the
complexity of GitHub's API object structure which can be large.