Updated structs to no longer use keyword arguments flag
Necessary to drop support for this flag since it'll most likely be removed entirely in future versions of Ruby. This also means the implementation was updated to splat arguments when creating new instances. In some cases, the `.for` method needed to be introduced because the parameter signature for keyword arguments for `.new` is different than `.for` due to being written in C instead of Ruby. Example:

    Ghub::Models::BooleanLink.method(:new).parameters  # [[:rest]]
    Ghub::Models::BooleanLink.method(:for).parameters  # [[:keyrest, :**]]

Milestone: minor